Inflation Expectations survey of households (IESH)  data
The data used in this project is related to the Inflation Expectations of people in India with respect to years 2021 and 2022. 
Website link: https://dbie.rbi.org.in/DBIE/dbie.rbi?site=unitLevelData
Metadata is available on the website for clear understanding of the various parameters that were used in the survey.